requirements_minimal.txt
# 🚀 Cruz 架構 - 最小化版本 (避免 Railway 4GB 限制)
Flask==2.3.3
Werkzeug==2.3.7
line-bot-sdk==3.9.0

# 🧠 AI 對話 - Gemini API
google-generativeai==0.8.0

# 🔊 語音合成
requests==2.31.0

# 🛠️ 基礎工具
python-dotenv==0.21.1
gunicorn==20.1.0

# 🎤 使用輕量 Whisper (避免重 PyTorch)
# 暫時移除 Whisper，先讓基本功能上線
# openai-whisper 包含了巨大的 PyTorch + CUDA (5GB+)
# 後續可考慮 whisper-jax 或其他輕量替代方案